+++ NEWS.md
@@ -7,6 +7,11 @@
 * function to transform existing surveys in model outputs for comparisons
 * automatic integration of Paracou and Guyafor data into model outputs
 
+# TROLL 0.1.9009
+
+* trait sensitivity analysis > sensitivity.Rmd
+* `compareplot` function for sensitivity analysis
+
 # TROLL 0.1.9008
 
 * fix init.R